Add MinGW-w64 cross-compilation support for Windows (Vulkan backend)

+++ lib/sdl/shader.cpp
@@ -56,9 +56,9 @@ namespace sdl
                 throw error("Failed to create shader");
             }
 
-            SDL_LogDebug(SDL_LOG_CATEGORY_APPLICATION, "Shader created: %s, %zu bytes",
+            SDL_LogDebug(SDL_LOG_CATEGORY_APPLICATION, "Shader created: %s, %lu bytes",
                          stage == SDL_GPU_SHADERSTAGE_VERTEX ? "vertex" : "fragment",
-                         code_size);
+                         (unsigned long)code_size);
         }
 
         ~impl() noexcept
